Pixelated shader fixed.
The pixelated shader is now fixed for the Tokyo level. This is done by simply modifying the logic of stencil layers, i.e., we now set every actor in the scene to have a layer stencil of 1, while our main character's stencil = 0. Now, in the blueprint of the pixelated material, we set a conditional statement such that if the stencil value >= 1, we'll use the shader, otherwise don't.