Dialogue Space Polished #1
Ideally, we want:
For all dialogue lines:
1. if it's the beginning of the line and
2. the sequence is lesser than that of the current one, and
3. it has not been displayed
-> Display those lines first before the current lines.

However, this is kinda difficult to implement as the data table is, by default, write only (it seems that we need to create another interface to deal with these using struct array instead, which is cumbersome!).

Instead, we make the dialogue pop up automatically when overlapping with the actor ->, i.e., the player won't miss any dialogue lines!